+++ src/CurrencyPrices.php
@@ -20,6 +20,7 @@ use kuriousagency\commerce\currencyprices\adjusters\Discount;
 use kuriousagency\commerce\currencyprices\twigextensions\CurrencyPricesTwigExtension;
 use kuriousagency\commerce\currencyprices\assetbundles\currencyprices\CurrencyPricesAsset;
 use kuriousagency\commerce\currencyprices\fields\CurrencyField;
+use kuriousagency\commerce\currencyprices\models\ShippingMethod as CurrencyPriceShippingMethod;
 
 use Craft;
 use craft\base\Plugin;
@@ -41,6 +42,9 @@ use craft\commerce\services\LineItems;
 use craft\commerce\events\ProcessPaymentEvent;
 use craft\commerce\services\Payments;
 use craft\commerce\services\OrderAdjustments;
+use craft\commerce\services\ShippingMethods;
+use craft\commerce\models\ShippingMethod;
+use craft\commerce\events\RegisterAvailableShippingMethodsEvent;
 
 use yii\base\Event;
 use craft\db\ActiveRecord;
@@ -302,14 +306,32 @@ class CurrencyPrices extends Plugin
 		Event::on(OrderAdjustments::class, OrderAdjustments::EVENT_REGISTER_ORDER_ADJUSTERS, function(RegisterComponentTypesEvent $e) {
 			foreach ($e->types as $key => $type)
 			{
-				if ($type == 'craft\\commerce\\adjusters\\Shipping') {
-					$e->types[$key] = Shipping::class;
-				}
+				// if ($type == 'craft\\commerce\\adjusters\\Shipping') {
+				// 	$e->types[$key] = Shipping::class;
+				// }
 				if ($type == 'craft\\commerce\\adjusters\\Discount') {
 					$e->types[$key] = Discount::class;
 				}
 			}
 			//Craft::dd($e->types);
+		});
+
+		Event::on(ShippingMethods::class, ShippingMethods::EVENT_REGISTER_AVAILABLE_SHIPPING_METHODS, function(RegisterAvailableShippingMethodsEvent $e) {
+
+			$order = $e->order;
+
+			foreach($e->shippingMethods as $key=>$method) {
+
+				if (get_class($method) === ShippingMethod::class) {
+					unset($e->shippingMethods[$key]);
+				}
+				
+				$newMethod = new CurrencyPriceShippingMethod($method);
+				$newMethod->order = $order;
+				$e->shippingMethods[] = $newMethod; 
+
+			}
+
 		});
 
 		/*Event::on(Discount::class, Discount::EVENT_AFTER_DISCOUNT_ADJUSTMENTS_CREATED, function(DiscountAdjustmentsEvent $e) {